fix(installer): Install agent before adding the user to additional groups
